Update to legible training

+++ legibility/src/dl_algos/q_networks.py
@@ -14,7 +14,7 @@ class QNetwork(nn.Module):
 	def __call__(self, x_orig: jnp.ndarray) -> jnp.ndarray:
 		x = jnp.array(x_orig)
 		for i in range(self.num_layers):
-			x = self.activation_function(nn.Dense(self.layer_sizes[i])(x))
+			x = self.activation_function(nn.Dense(self.layer_sizes[i], dtype=jnp.float32)(x))
 		return nn.Dense(self.action_dim)(x)
 
 
@@ -28,7 +28,7 @@ class DuelingQNetwork(nn.Module):
 	def __call__(self, x_orig: jnp.ndarray) -> jnp.ndarray:
 		x = jnp.array(x_orig)
 		for i in range(self.num_layers):
-			x = self.activation_function(nn.Dense(self.layer_sizes[i])(x))
+			x = self.activation_function(nn.Dense(self.layer_sizes[i], dtype=jnp.float32)(x))
 		a = nn.Dense(self.action_dim)(x)
 		v = nn.Dense(1)(x)
 		return v + (a - a.mean())
@@ -48,11 +48,11 @@ class CNNQNetwork(nn.Module):
 	def __call__(self, x_orig: jnp.ndarray) -> jnp.ndarray:
 		x = x_orig
 		for i in range(self.num_conv_layers):
-			x = self.activation_function(nn.Conv(self.cnn_size[i], kernel_size=self.cnn_kernel[i], padding='SAME')(x))
+			x = self.activation_function(nn.Conv(self.cnn_size[i], kernel_size=self.cnn_kernel[i], padding='SAME', dtype=jnp.float32)(x))
 			x = nn.max_pool(x, window_shape=self.pool_window[i], padding='VALID')
 		x = x.reshape((x.shape[0], -1))
 		for i in range(self.num_linear_layers):
-			x = self.activation_function(nn.Dense(self.layer_sizes[i])(x))
+			x = self.activation_function(nn.Dense(self.layer_sizes[i], dtype=jnp.float32)(x))
 		return nn.Dense(self.action_dim)(x)
 
 
@@ -70,11 +70,11 @@ class CNNDuelingQNetwork(nn.Module):
 	def __call__(self, x_orig: jnp.ndarray) -> jnp.ndarray:
 		x = x_orig
 		for i in range(self.num_conv_layers):
-			x = self.activation_function(nn.Conv(self.cnn_size[i], kernel_size=self.cnn_kernel[i], padding='SAME')(x))
+			x = self.activation_function(nn.Conv(self.cnn_size[i], kernel_size=self.cnn_kernel[i], padding='SAME', dtype=jnp.float32)(x))
 			x = nn.max_pool(x, window_shape=self.pool_window[i], padding='VALID')
 		x = x.reshape((x.shape[0], -1))
 		for i in range(self.num_linear_layers):
-			x = self.activation_function(nn.Dense(self.layer_sizes[i])(x))
+			x = self.activation_function(nn.Dense(self.layer_sizes[i], dtype=jnp.float32)(x))
 		a = nn.Dense(self.action_dim)(x)
 		v = nn.Dense(1)(x)
 		return v + (a - a.mean())
@@ -98,11 +98,11 @@ class MultiObsCNNDuelingQNetwork(nn.Module):
 		for idx in range(self.num_obs):
 			x_temp = x_orig[:, idx]
 			for i in range(self.num_conv_layers):
-				x_temp = self.activation_function(nn.Conv(self.cnn_size[i], kernel_size=self.cnn_kernel[i], padding='SAME')(x_temp))
+				x_temp = self.activation_function(nn.Conv(self.cnn_size[i], kernel_size=self.cnn_kernel[i], padding='SAME', dtype=jnp.float32)(x_temp))
 				x_temp = nn.max_pool(x_temp, window_shape=self.pool_window[i], padding='VALID')
 			x = jnp.append(x, x_temp).reshape((x_orig.shape[0], -1))
 		for i in range(self.num_linear_layers):
-			x = self.activation_function(nn.Dense(self.layer_sizes[i])(x))
+			x = self.activation_function(nn.Dense(self.layer_sizes[i], dtype=jnp.float32)(x))
 		a = nn.Dense(self.action_dim)(x)
 		v = nn.Dense(1)(x)
 		return v + (a - a.mean())
